+++ src/vehicle.cpp
@@ -3703,7 +3703,7 @@ void vehicle::fire_turret (int p, bool burst)
     if (!gun) {
         return;
     }
-    int charges = burst? gun->burst : 1;
+    long charges = burst? gun->burst : 1;
     std::string whoosh = "";
     if (!charges)
         charges = 1;
@@ -3768,7 +3768,7 @@ void vehicle::fire_turret (int p, bool burst)
     }
 }
 
-bool vehicle::fire_turret_internal (int p, it_gun &gun, it_ammo &ammo, int charges, const std::string &extra_sound)
+bool vehicle::fire_turret_internal (int p, it_gun &gun, it_ammo &ammo, long charges, const std::string &extra_sound)
 {
     int x = global_x() + parts[p].precalc_dx[0];
     int y = global_y() + parts[p].precalc_dy[0];
